quicknotes-app/requirements.txt
Flask==2.3.3
PyMySQL==1.1.1
cryptography==42.0.8
python-dotenv==1.0.0
